requirements.txt
agate==1.9.1
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.13.0
asgi-csrf==0.11
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
daff==1.4.2
datasette==0.65.2
dbt-adapters==1.23.0
dbt-common==1.38.0
dbt-core==1.11.9
dbt-duckdb==1.10.1
dbt-extractor==0.6.0
dbt-protos==1.0.491
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
defusedxml==0.7.1
Django==6.0.5
django-cors-headers==4.9.0
djangorestframework==3.17.1
Faker==40.18.0
flexcache==0.3
flexparser==0.4
fonttools==4.63.0
git-filter-repo==2.47.0
gunicorn==26.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hupper==1.12.1
idna==3.13
importlib_metadata==8.9.0
isodate==0.7.2
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jira==3.10.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
leather==0.4.1
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.9
mdurl==0.1.2
mergedeep==1.3.4
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
orderly-set==5.5.0
packaging==26.2
parsedatetime==2.6
pathspec==0.12.1
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
protobuf==6.33.6
psycopg2-binary==2.9.12
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
scipy @ file:///home/builder/.termux-build/python-scipy/src/dist/scipy-1.17.1-cp313-cp313-android_aarch64.whl#sha256=983a3b6ea280209902b8cc8c5c47d881237522224142999ddf2c8e084ffb2db3
setuptools==82.0.1
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.5
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
uvicorn==0.46.0
wheel==0.47.0
zipp==3.23.1
